Safari-609.1.11.1/LayoutTests/accessibility/mac/heading-clickpoint-expected.txt

Small heading

A heading that is just a little bit wider

Small heading Small heading

This tests that the click points for headings should use the only child's clickpoint instead of taking the mid point (which may be on nothing).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS heading1.clickPointX is heading1.childAtIndex(0).clickPointX
PASS heading1.clickPointY is heading1.childAtIndex(0).clickPointY
PASS heading2.clickPointX is heading2.childAtIndex(0).clickPointX
PASS heading2.clickPointY is heading2.childAtIndex(0).clickPointY
PASS heading3.clickPointX == heading3.childAtIndex(0).clickPointX is false
PASS successfullyParsed is true

TEST COMPLETE
